requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0
google-api-python-client>=2.70.0
google-auth>=2.16.0
google-auth-httplib2>=0.1.0
PyYAML>=6.0
